Restructured code to append documents
In order to resolve a PMD error, I moved the file IO and document
append code to its own function so that it can be easily called for
each document that we want merged. I also removed one of the three
input files for simplicity's sake, since the third was redundant.