Create a sample to create PDFs from images
This sample will show how to create a PDF from a JPG, PNG, BMP, or
GIF image. The images I have added seem to be lacking proper metadata
for us to determine their DPI. The question is whether the code is
looking at the metadata wrong, or if these images are just special.
Note: a JPG was downloaded from the internet, then saved in different
formats using IrfanView. For what it's worth.