Send job to default printer 'silently'
Don't present a print dialog when printing, just send it directly to
the default printer.

Note that this fixes a subtle bug: we requested the system default
printer and use its default resolution to rasterize the page...then
presented a dialog where the user might choose a *different* printer
with a different resolution.

By always printing silently, we guarantee that we are using the
correct resolution.