Implement a fake PrinterJob to intercept print requests
We use mocking to inject our own 'fake' PrinterJob that will intercept
print requests and render them to a private BufferedImage. This is the
minimal logic required to make this work; still to-do is to save the
buffered page images to disk and checksum their results.

This also cleans up earlier mocks that are no longer required (and
weren't working as desired anyway).