DocumentSigning sample : update input file
This still isn’t perfect, it at least looks a little
nicer though as it has the Datalogics name at the
top. I need to put some text in it though to have
it look really nice.

The input file also only has one signature field and
it has a title of “Approver” rather than the default
name that Acrobat provides so the test has been
updated as well.