Do not manually append ".pdf" to signing output
According to Jeremy, this is a relic of the fact that the
output files were being enumerated and thus needed to sneak
their number in before the .pdf. However, since that is no
longer the case it no longer makes sense to request the user
give us a path to a PDF to sign and a path to put it, and then add ".pdf" to the given output path.