Add an integration test to run all the main methods from the jar.
Integration tests happen during the integration-test and verify phases
in Maven. They're designed to run the built package in a constructed
environment.

The reflections package is used as a helper to find all the classes
in the current package, and then all of the main methods in each
class. A parameterized test is then created and run for each main
method.

This only tests that the samples will work from the jar. The unit
tests verify that the sample methods work correctly.