POM: Run the integration tests in the integration-tests profile
To run the integration tests, do

mvn verify -P integration-tests

The Jenkins builders will be set up to run integration tests.

By default, they won't take up developers' time when run from the
command line.

The integration tests are run with the classpath set to the
packaged jar. To ensure that the integration tests are in a clean
environment, the target/classes directory is deleted, and the tests
are run with the current directory set to
target/integration-test-outputs.

Note that this means that running the integration tests may mean that
you have to rebuild the project in Eclipse before things run again.