Add test suites to run AllIntegrationTests or AllUnitTests
There is no project setting to keep Eclipse from running all the tests.

Therefore, two test suites are provided, AllIntegrationTests and
AllUnitTests.

Developers will normally want to select AllUnitTests and run those
as JUnit tests. This will run just the unit tests, and not the
integration tests.

Note that running the whole package as a JUnit test will run all the
tests *twice*, once individually, and once as a suite.

Remembering to add tests to the test suites is hard, so unit-toolbox
is used to create test suites via wildcards based on the test class
names.