Refactor the redact and sanitize sample to follow new input/output pattern.
The sample's main method now takes care of input and output only. It
processes the args[] array and creates input and output URLs out of the
arguments. If nothing is passed in, it uses default values to create the
URLs.

The tests now call a redactAndSanitize method, instead of the main method.

The "CanSanitize" test now targets the sanitizeMethod more specifically.