Add includes to pom
In order to get our samples ready for distribution, I've added a long
list of includes in the pom.xml for the samples project. This will
ensure that any other files added to the repo will not be included in
the Maven build.

To test this, I added a new source file in my workspace but didn't
add it to the include list, then ran mvn install. When I extracted
the jar, this new class was not there, as expected.