Reworked the CreatePdfFromImage sample.
The sample now takes care of input and output in the main method.

There is now a method that takes a list of input image URLs and creates
PDF documents based on that. Tests were also added for this method.

Another method added to save a document when it creates it.

All new methods call createPdfFromImage to do the actual work.