Fixed test inputs, set copy filter to false in pom
Some tests don't take inputs, but I had been assuming they did. Fixed
that. Also, in the file copy plugin, the filtering option was set to
true because I assumed that was a filter for file names. Still not
totally sure what it is, but having it on messes PDFs up bad.