Added a check for another XML element to FillForm
This issue was brought up by SF case 38873. The customer had a PDF form
document and a sample program that produced what appeared to be blank
output. Running it through the FillForm sample with a data file I
created seemed to reproduce it somewhat. Ultimately, it turned out the
data file was missing an XML element that the XFAService required.
Now the sample will check for that element before filling the form.
The customer can now use this sample to write their own code.