Add LogRecordListCollector class to utility package
This adds a utility for gathering log messages, to be used during
testing.

It duplicates a class used in the Talkeetna tests. Due to the structure
of our repos, the implementation in Talkeetna tests cannot be used;
a solution to this issue is outside of the scope for this user story.