lite: Supply the super-pom in the customer distribution
This fixes a chicken-and-egg problem where the local repo can't be found via
project variables before the super-pom is read.

The super-pom is now included as a subdirectory of the shipped samples. The
maven-dependency-plugin is used to copy it, and the copy-rename-maven-plugin
moves it into place in the distribution.

When building from the customer package, the super-pom is found via the
relativePath.

When building otherwise, the super-pom will be found with a repo search.